Make indentation-engine use regular expressions.
Previous implementation indented on the first word of the line only, but
that limited the ability wsd-mode had to support the full wsd website
script-grammer.

This change also introduces support for new keywords and highlighting of
those.

Test-suite has been expanded to cover all ned indentation logic.

This commit addresses the following issue:

https://github.com/josteink/wsd-mode/issues/26
https://github.com/josteink/wsd-mode/issues/27
https://github.com/josteink/wsd-mode/issues/21