Encode message so that it can be safely transported in the url
The url-encode-url function does not escape all characters that cause problems
when the message is sent as a url encoded form parameter. I discovered this
when some of my diagrams failed because I had comments with an & in them. I
believe url-encode-url is intended for things that are already URIs, whereas
when a parameter value needs to be encoded as here url-hexify-string is the
correct function to use
(https://www.gnu.org/software/emacs/manual/html_node/url/URI-Encoding.html).

I suppose wsd-encode can be inlined now since it just wraps url-hexify-string,
but I'll leave that decision to the maintainer.